Api web

+++ API/SignalR/ChatHub.cs
@@ -10,11 +10,10 @@ namespace API.SignalR
     [Authorize]
     public class ChatHub : Hub
     {
-        //IMapper _mapper;
-        IHubContext<PresenceHub> _presenceHub;
-        PresenceTracker _presenceTracker;
-        IUnitOfWork _unitOfWork;
-        UserShareScreenTracker _shareScreenTracker;
+        private IHubContext<PresenceHub> _presenceHub;
+        private PresenceTracker _presenceTracker;
+        private IUnitOfWork _unitOfWork;
+        private UserShareScreenTracker _shareScreenTracker;
 
         public ChatHub(IUnitOfWork unitOfWork, UserShareScreenTracker shareScreenTracker, PresenceTracker presenceTracker, IHubContext<PresenceHub> presenceHub)
         {
@@ -28,45 +27,58 @@ namespace API.SignalR
         public override async Task OnConnectedAsync()
         {
             var httpContext = Context.GetHttpContext();
-            var roomId = httpContext.Request.Query["roomId"].ToString();
-            var roomIdInt = int.Parse(roomId);
+
+            if (!Guid.TryParse(httpContext?.Request.Query["roomId"].ToString(), out var roomId))
+                throw new HubException("Invalid query param");
+
+            if (Context.User == null)
+                throw new HubException("401");
+
             var userId = Context.User.GetUserId();
             var displayName = Context.User.GetDisplayName();
-            var userConnection = new UserConnectionInfo(userId, displayName, roomIdInt);
+
+            var userConnection = new UserConnectionInfo(userId, displayName, roomId);
 
             await _presenceTracker.UserConnected(userConnection, Context.ConnectionId);
 
-            await Groups.AddToGroupAsync(Context.ConnectionId, roomId);//khi user click vao room se join vao
-            await AddConnectionToGroup(roomIdInt); // luu db DbSet<Connection> de khi disconnect biet
+            await Groups.AddToGroupAsync(Context.ConnectionId, roomId.ToString());//khi user click vao room se join vao
+            await AddConnectionToGroup(roomId); // luu db DbSet<Connection> de khi disconnect biet
 
             //var usersOnline = await _unitOfWork.UserRepository.GetUsersOnlineAsync(currentUsers);
             var oneUserOnline = await _unitOfWork.UserRepository.GetMemberAsync(userId);
-            await Clients.Group(roomId).SendAsync("UserOnlineInGroup", oneUserOnline);
+            await Clients.Group(roomId.ToString()).SendAsync("UserOnlineInGroup", oneUserOnline);
 
-            var currentUsers = await _presenceTracker.GetOnlineUsers(roomIdInt);
-            await _unitOfWork.RoomRepository.UpdateCountMember(roomIdInt, currentUsers.Length);
+            var currentUsers = await _presenceTracker.GetOnlineUsers(roomId);
+            await _unitOfWork.RoomRepository.UpdateCountMember(roomId, currentUsers.Length);
             await _unitOfWork.Complete();
 
             var currentConnections = await _presenceTracker.GetConnectionsForUser(userConnection);
             await _presenceHub.Clients.AllExcept(currentConnections).SendAsync("CountMemberInGroup",
-                   new { roomId = roomIdInt, countMember = currentUsers.Length });
+                   new { roomId = roomId, countMember = currentUsers.Length });
 
             //share screen user vao sau cung
-            var userIsSharing = await _shareScreenTracker.GetUserIsSharing(roomIdInt);
+            var userIsSharing = await _shareScreenTracker.GetUserIsSharing(roomId);
             if (userIsSharing != null)
             {
                 var currentBeginConnectionsUser = await _presenceTracker.GetConnectionsForUser(userIsSharing);
-                if (currentBeginConnectionsUser.Count > 0)
+                if (currentBeginConnectionsUser?.Count > 0)
                     await Clients.Clients(currentBeginConnectionsUser).SendAsync("OnShareScreenLastUser", new { userIdTo = userId, isShare = true });
                 await Clients.Caller.SendAsync("OnUserIsSharing", userIsSharing.DisplayName);
             }
         }
 
-        public override async Task OnDisconnectedAsync(Exception exception)
+        public override async Task OnDisconnectedAsync(Exception? exception)
         {
+            if (Context.User == null)
+                throw new HubException("401");
+
             var userId = Context.User.GetUserId();
             var displayName = Context.User.GetDisplayName();
             var group = await RemoveConnectionFromGroup();
+
+            if (group == null)
+                throw new HubException("Room doesn't exists");
+
             var isOffline = await _presenceTracker.UserDisconnected(new UserConnectionInfo(userId, displayName, group.RoomId), Context.ConnectionId);
 
             await _shareScreenTracker.DisconnectedByUser(userId, group.RoomId);
@@ -83,29 +95,54 @@ namespace API.SignalR
 
                 await _presenceHub.Clients.All.SendAsync("CountMemberInGroup",
                        new { roomId = group.RoomId, countMember = currentUsers.Length });
+
+                /*if (currentUsers.Length < 1 || Context.User.IsInRole("Host") || Context.User.IsInRole("Admin"))
+                    await LockdownRoom(group.RoomId);*/
             }
             await base.OnDisconnectedAsync(exception);
         }
 
         public async Task SendMessage(CreateMessageDto createMessageDto)
         {
+            if (Context.User == null)
+                throw new HubException("401");
+
             var userId = Context.User.GetUserId();
             var displayName = Context.User.GetDisplayName();
             var group = await _unitOfWork.RoomRepository.GetRoomForConnection(Context.ConnectionId);
 
+            if (group == null)
+                throw new HubException("group == null");
+
+            if (group.BlockedChat)
+                throw new HubException("Chat has been blocked by host");
+
+            var message = new MessageDto
+            {
+                SenderUserID = userId,
+                SenderDisplayName = displayName,
+                Content = createMessageDto.Content,
+                MessageSent = DateTime.UtcNow
+            };
+            //Luu message vao db
+            //code here
+            //send meaasge to group
+            await Clients.Group(group.RoomId.ToString()).SendAsync("NewMessage", message);
+        }
+
+        [Authorize(Roles = "Admin,Host")]
+        public async Task BlockChat(bool block)
+        {
+            if (Context.User == null)
+                throw new HubException("401");
+
+            var group = await _unitOfWork.RoomRepository.GetRoomForConnection(Context.ConnectionId);
+
             if (group != null)
             {
-                var message = new MessageDto
-                {
-                    SenderUserID = userId,
-                    SenderDisplayName = displayName,
-                    Content = createMessageDto.Content,
-                    MessageSent = DateTime.Now
-                };
-                //Luu message vao db
-                //code here
-                //send meaasge to group
-                await Clients.Group(group.RoomId.ToString()).SendAsync("NewMessage", message);
+                await _unitOfWork.RoomRepository.UpdateBlockChat(group.RoomId, block);
+                await _unitOfWork.Complete();
+                await Clients.Group(group.RoomId.ToString()).SendAsync("OnBlockChat", new { block });
             }
         }
 
@@ -114,9 +151,33 @@ namespace API.SignalR
             var group = await _unitOfWork.RoomRepository.GetRoomForConnection(Context.ConnectionId);
             if (group != null)
             {
+                if (Context.User == null)
+                    throw new HubException("401");
+
                 await Clients.Group(group.RoomId.ToString()).SendAsync("OnMuteMicro", new
                 {
                     userId = Context.User.GetUserId(),
+                    mute = muteMicro
+                });
+            }
+            else
+            {
+                throw new HubException("group == null");
+            }
+        }
+
+        [Authorize(Roles = "Admin,Host")]
+        public async Task MuteAllMicro(Guid userId, bool muteMicro)
+        {
+            var group = await _unitOfWork.RoomRepository.GetRoomForConnection(Context.ConnectionId);
+            if (group != null)
+            {
+                if (!group.Connections.Any(item => item.UserID == userId))
+                    throw new HubException("user_id not in current group");
+
+                await Clients.Group(group.RoomId.ToString()).SendAsync("OnMuteAllMicro", new
+                {
+                    userId = userId,
                     mute = muteMicro
                 });
             }
@@ -131,6 +192,9 @@ namespace API.SignalR
             var group = await _unitOfWork.RoomRepository.GetRoomForConnection(Context.ConnectionId);
             if (group != null)
             {
+                if (Context.User == null)
+                    throw new HubException("401");
+
                 await Clients.Group(group.RoomId.ToString()).SendAsync("OnMuteCamera", new
                 {
                     userId = Context.User.GetUserId(),
@@ -143,13 +207,16 @@ namespace API.SignalR
             }
         }
 
-        public async Task ShareScreen(int roomid, bool isShareScreen)
+        public async Task ShareScreen(Guid roomid, bool isShareScreen)
         {
+            if (Context.User == null)
+                throw new HubException("401");
+
             var userConnection = new UserConnectionInfo(Context.User.GetUserId(), Context.User.GetDisplayName(), roomid);
             if (isShareScreen)//true is doing share
             {
                 await _shareScreenTracker.UserConnectedToShareScreen(userConnection);
-                await Clients.Group(roomid.ToString()).SendAsync("OnUserIsSharing", Context.User.GetUsername());
+                await Clients.Group(roomid.ToString()).SendAsync("OnUserIsSharing", Context.User.GetDisplayName());
             }
             else
             {
@@ -159,17 +226,44 @@ namespace API.SignalR
             //var group = await _unitOfWork.RoomRepository.GetRoomForConnection(Context.ConnectionId);
         }
 
-        public async Task ShareScreenToUser(int roomid, Guid userId, bool isShare)
+        public async Task ShareScreenToUser(Guid roomid, Guid userId, bool isShare)
         {
             var currentBeginConnectionsUser = await _presenceTracker.GetConnectionsForUser(new UserConnectionInfo(userId, string.Empty, roomid));
-            if (currentBeginConnectionsUser.Count > 0)
+            if (currentBeginConnectionsUser?.Count > 0)
                 await Clients.Clients(currentBeginConnectionsUser).SendAsync("OnShareScreen", isShare);
         }
 
-        private async Task<Room> RemoveConnectionFromGroup()
+        [Authorize(Roles = "Admin,Host")]
+        public async Task KickMember(Guid roomId, Guid userId)
+        {
+            var connections = await _presenceTracker.GetConnectionsForUser(new UserConnectionInfo(userId, string.Empty, roomId));
+            if (connections != null && connections.Count > 0)
+            {
+                await Clients.Users(userId.ToString()).SendAsync("OnIsKicked", new { userId, roomId });
+                await Task.WhenAll(connections.Select(cid => Groups.RemoveFromGroupAsync(cid, roomId.ToString())));
+                var u = await _unitOfWork.UserRepository.UpdateLocked(userId);
+                var temp = await _unitOfWork.UserRepository.GetMemberAsync(userId);
+                await Clients.Group(roomId.ToString()).SendAsync("UserOfflineInGroup", temp);
+
+                var currentUsers = await _presenceTracker.GetOnlineUsers(roomId);
+
+                await _unitOfWork.RoomRepository.UpdateCountMember(roomId, currentUsers.Length);
+                await _unitOfWork.Complete();
+
+                await _presenceHub.Clients.All.SendAsync("CountMemberInGroup",
+                       new { roomId = roomId, countMember = currentUsers.Length });
+            }
+        }
+
+        #region Private
+        private async Task<Room?> RemoveConnectionFromGroup()
         {
             var group = await _unitOfWork.RoomRepository.GetRoomForConnection(Context.ConnectionId);
-            var connection = group.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+            var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+
+            if (connection == null)
+                throw new HubException("Room doesn't exists");
+
             _unitOfWork.RoomRepository.RemoveConnection(connection);
 
             if (await _unitOfWork.Complete()) return group;
@@ -177,8 +271,11 @@ namespace API.SignalR
             throw new HubException("Fail to remove connection from room");
         }
 
-        private async Task<Room> AddConnectionToGroup(int roomId)
+        private async Task<Room?> AddConnectionToGroup(Guid roomId)
         {
+            if (Context.User == null)
+                throw new HubException("401");
+
             var group = await _unitOfWork.RoomRepository.GetRoomById(roomId);
             var connection = new Connection(Context.ConnectionId, Context.User.GetUserId());
             if (group != null)
@@ -190,5 +287,24 @@ namespace API.SignalR
 
             throw new HubException("Failed to add connection to room");
         }
+
+        private async Task LockdownRoom(Guid roomId)
+        {
+            await Clients.Group(roomId.ToString()).SendAsync("LockdownRoom");
+            var connectionsInRoom = await _unitOfWork.RoomRepository.GetRoomById(roomId);
+            await _presenceTracker.RemoveAllConnectionByRoomID(roomId);
+            if (connectionsInRoom != null)
+            {
+                await Task.WhenAll(connectionsInRoom.Connections.Select(item => Groups.RemoveFromGroupAsync(item.ConnectionId, roomId.ToString())));
+                _unitOfWork.RoomRepository.RemoveConnections(connectionsInRoom.Connections);
+                await _unitOfWork.RoomRepository.DeleteRoom(roomId);
+                foreach (var user in connectionsInRoom.Connections)
+                {
+                    await _unitOfWork.UserRepository.UpdateLocked(user.UserID);
+                }
+            }
+            await _unitOfWork.Complete();
+        }
+        #endregion
     }
 }
